Linking

+++ server/app/Services/ElectionsService.php
@@ -38,6 +38,7 @@ class ElectionsService
         $addedElections->title = $request->title;
         $addedElections->description = $request->description;
         $addedElections->on_going = true;
+        $addedElections->save();
 
         return $addedElections;
     
@@ -55,8 +56,12 @@ class ElectionsService
             abort(403, 'Cannot Stop elections');
         }
 
-        $elections = Elections::FindOrFail($request->id);
-        $elections->delete();
+        $elections = Elections::Find($request->id);
+        if ($elections) {
+            $elections->delete();
+        } else {
+            return ["message" => "No elections found"];
+        }
 
         return $elections;
